requirements.txt
Flask>=2.0
Flask-SQLAlchemy
Flask-Login
Flask-Migrate
Flask-Mail
APScheduler
python-dotenv
passlib
